+++ CHANGELOG.md
@@ -7,3 +7,9 @@ Changelog
 ### 0.1.0
 - Updating `vagrant up` to include building dependencies on first load
 - Adding Gemfile and bower.json to handle additional dependencies
+
+### 0.1.1
+- Removing auto build on `vagrant up` due to failing GETs
+- Removing Ruby Sass dependancy in favour of libsass
+- Removing jquery.hammer from bower.json
+- Updating `<html>` conditional comments